requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
asgiref==3.8.1
    # via
    #   django
    #   django-htmx
black==24.4.2
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via black
colorama==0.4.6
    # via
    #   click
    #   pytest
cryptography==42.0.7
defusedxml==0.7.1
    # via python3-openid
django==5.0.6
    # via
    #   django-allauth
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-htmx
django-allauth==0.63.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-filter==24.2
django-htmx==1.17.3
django-widget-tweaks==1.5.0
factory-boy==3.3.0
faker==25.2.0
    # via factory-boy
idna==3.7
    # via requests
iniconfig==2.0.0
    # via pytest
mypy-extensions==1.0.0
    # via black
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.0
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.2.2
    # via black
pluggy==1.5.0
    # via pytest
pycparser==2.22
    # via cffi
pyjwt==2.8.0
pytest==8.2.1
    # via pytest-django
pytest-django==4.8.0
python-dateutil==2.9.0.post0
    # via faker
python3-openid==3.2.0
requests==2.31.0
    # via requests-oauthlib
requests-oauthlib==2.0.0
ruff==0.4.4
six==1.16.0
    # via python-dateutil
sqlparse==0.5.0
    # via
    #   django
    #   django-debug-toolbar
tomli==2.0.1
tzdata==2024.1
    # via django
urllib3==2.2.1
    # via requests
